refactor: 랜덤 유틸리티에서 서비스로 변경
