+++ src/main/java/lotto/service/RandomService.java
@@ -1,0 +1,8 @@
+package lotto.service;
+
+import java.util.List;
+
+public interface RandomService {
+
+    List<List<Integer>> generateRandomNumbersList(int count, int maxSize, int maxNumber);
+}
