+++ src/main/java/lotto/service/RandomServiceImpl.java
@@ -1,4 +1,4 @@
-package lotto.utils;
+package lotto.service;
 
 import java.util.ArrayList;
 import java.util.HashSet;
@@ -7,18 +7,15 @@ import java.util.Random;
 import java.util.Set;
 import java.util.stream.IntStream;
 
-public final class RandomNumberUtils {
+public class RandomServiceImpl implements RandomService {
 
     private static final Random random = new Random();
 
-    private RandomNumberUtils() {
-    }
-
-    private static int generateRandomNumber(int maxNumber) {
+    private int generateRandomNumber(int maxNumber) {
         return random.nextInt(maxNumber) + 1;
     }
 
-    public static List<Integer> generateRandomNumbers(int maxSize, int maxNumber) {
+    public List<Integer> generateRandomNumbers(int maxSize, int maxNumber) {
         final Set<Integer> numberSet = new HashSet<>();
 
         while (numberSet.size() < maxSize) {
@@ -28,7 +25,7 @@ public final class RandomNumberUtils {
         return new ArrayList<>(numberSet).stream().sorted().toList();
     }
 
-    public static List<List<Integer>> generateRandomNumbersList(int count, int maxSize, int maxNumber) {
+    public List<List<Integer>> generateRandomNumbersList(int count, int maxSize, int maxNumber) {
         return IntStream.range(0, count)
                 .mapToObj(i -> generateRandomNumbers(maxSize, maxNumber))
                 .toList();
